+++ src/editor/docs/Docs.svelte.d.ts
@@ -1,0 +1,2 @@
+import { SvelteComponent } from 'svelte';
+export default class Docs extends SvelteComponent<Record<string, never>> {}